+++ src/biz/bokhorst/xprivacy/XTelephonyRegistry.java
@@ -1,6 +1,7 @@
 package biz.bokhorst.xprivacy;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -10,19 +11,21 @@ import android.content.Context;
 import android.os.Binder;
 import android.os.Bundle;
 import android.os.IBinder;
+import android.os.IInterface;
+import android.os.Process;
 import android.telephony.CellInfo;
-import android.telephony.ServiceState;
-import android.telephony.SignalStrength;
 import android.util.Log;
 
-import com.android.internal.telephony.IPhoneStateListener;
-
+import de.robv.android.xposed.XC_MethodHook;
+import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XC_MethodHook.MethodHookParam;
 import static de.robv.android.xposed.XposedHelpers.findField;
 
 public class XTelephonyRegistry extends XHook {
 	private Methods mMethod;
-	private static final Map<IBinder, XIPhoneStateListener> mListener = new WeakHashMap<IBinder, XIPhoneStateListener>();
+	private static boolean mHookedLocation = false;
+	private static boolean mHookedPhone = false;
+	private static Map<IBinder, Integer> mListenerUid = new WeakHashMap<IBinder, Integer>();
 
 	private XTelephonyRegistry(Methods method, String restrictionName) {
 		super(restrictionName, method.name(), String.format("Srv.%s", method.name()));
@@ -60,45 +63,42 @@ public class XTelephonyRegistry extends XHook {
 	@Override
 	protected void before(MethodHookParam param) throws Throwable {
 		if (mMethod == Methods.listen) {
-			if (isRestricted(param))
-				replacePhoneStateListener(param);
-		} else
-			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
-	}
+			if (param.args.length > 2 && param.args[1] != null)
+				if (param.args[1].getClass().getName().startsWith("com.android.internal.telephony.IPhoneStateListener")) {
+					IBinder binder = ((IInterface) param.args[1]).asBinder();
+					int event = (Integer) param.args[2];
+					if (event == android.telephony.PhoneStateListener.LISTEN_NONE) {
+						// Remove listener
+						Util.log(this, Log.WARN, "remove");
+						synchronized (mListenerUid) {
+							if (mListenerUid.containsKey(binder))
+								mListenerUid.remove(binder);
+							else
+								Util.log(this, Log.WARN, "Remove: listener not found");
+						}
+					} else {
+						// Add listener
+						int uid = Binder.getCallingUid();
+						Util.log(this, Log.WARN, "put uid=" + uid);
+						synchronized (mListenerUid) {
+							mListenerUid.put(binder, uid);
+						}
 
-	private void replacePhoneStateListener(MethodHookParam param) throws Throwable {
-		if (param.args.length > 2 && param.args[1] != null)
-			if (param.args[1] instanceof IPhoneStateListener) {
-				int event = (Integer) param.args[2];
-				IPhoneStateListener listener = (IPhoneStateListener) param.args[1];
-				if (event == android.telephony.PhoneStateListener.LISTEN_NONE) {
-					// Remove
-					synchronized (mListener) {
-						XIPhoneStateListener xlistener = mListener.get(listener.asBinder());
-						if (xlistener == null)
-							Util.log(this, Log.WARN, "Unknown class=" + listener.getClass().getName() + " method="
-									+ param.method + " count=" + mListener.size() + " uid=" + Binder.getCallingUid());
-						else {
-							param.args[1] = xlistener;
-							mListener.remove(listener.asBinder());
-							Util.log(this, Log.INFO, "Removed class=" + listener.getClass().getName() + " method="
-									+ param.method + " count=" + mListener.size() + " uid=" + Binder.getCallingUid());
+						// Hook
+						if (!mHookedLocation && "location".equals(getRestrictionName())) {
+							hookOn(param, "onCellLocationChanged");
+							hookOn(param, "onCellInfoChanged");
+							mHookedLocation = true;
 						}
-					}
-				} else {
-					// Replace
-					if (!(param.args[1] instanceof XIPhoneStateListener)) {
-						XIPhoneStateListener xListener = new XIPhoneStateListener(listener);
-						synchronized (mListener) {
-							mListener.put(listener.asBinder(), xListener);
-							Util.log(this, Log.INFO, "Added class=" + listener.getClass().getName() + " method="
-									+ param.method + " count=" + mListener.size() + " uid=" + Binder.getCallingUid());
+
+						if (!mHookedPhone && "phone".equals(getRestrictionName())) {
+							hookOn(param, "onCallStateChanged");
+							mHookedPhone = true;
 						}
-						param.args[1] = xListener;
 					}
 				}
-			} else
-				Util.log(this, Log.WARN, "Unexpected method=" + param.method + " uid=" + Binder.getCallingUid());
+		} else
+			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
 	}
 
 	@Override
@@ -106,88 +106,71 @@ public class XTelephonyRegistry extends XHook {
 		// Do nothing
 	}
 
-	@Override
-	protected boolean isRestricted(MethodHookParam param) throws Throwable {
-		Context context = null;
-
+	private void hookOn(final MethodHookParam param, final String methodName) {
 		try {
-			Field fieldContext = findField(param.thisObject.getClass(), "mContext");
-			context = (Context) fieldContext.get(param.thisObject);
-		} catch (NoSuchFieldError ex) {
+			// void onCallStateChanged(int state, String incomingNumber)
+			// void onCellLocationChanged(Bundle location)
+			// void onCellInfoChanged(List<CellInfo> cellInfo)
+			Method on = null;
+			if ("onCallStateChanged".equals(methodName))
+				on = param.args[1].getClass().getDeclaredMethod(methodName, int.class, String.class);
+			if ("onCellLocationChanged".equals(methodName))
+				on = param.args[1].getClass().getDeclaredMethod(methodName, Bundle.class);
+			if ("onCellInfoChanged".equals(methodName))
+				on = param.args[1].getClass().getDeclaredMethod(methodName, List.class);
+			XposedBridge.hookMethod(on, new XC_MethodHook() {
+				@Override
+				protected void beforeHookedMethod(MethodHookParam onparam) throws Throwable {
+					// Get uid
+					int uid = 0;
+					IBinder binder = ((IInterface) onparam.thisObject).asBinder();
+					synchronized (mListenerUid) {
+						if (mListenerUid.containsKey(binder))
+							uid = mListenerUid.get(binder);
+						else
+							Util.log(XTelephonyRegistry.this, Log.WARN, "Get: listener not found");
+					}
+					Util.log(XTelephonyRegistry.this, Log.WARN, "get uid=" + uid);
+
+					// Restrict
+					if (uid > 0) {
+						Util.log(XTelephonyRegistry.this, Log.WARN, methodName);
+						if (methodName.equals("onCallStateChanged")) {
+							if (onparam.args.length > 1 && onparam.args[1] != null)
+								if (isRestricted(param, uid))
+									onparam.args[1] = PrivacyManager.getDefacedProp(uid, "PhoneNumber");
+						} else if (methodName.equals("onCellLocationChanged")) {
+							if (onparam.args.length > 0 && onparam.args[0] != null)
+								if (isRestricted(param, uid))
+									onparam.args[0] = new Bundle();
+						} else if (methodName.equals("onCellInfoChanged")) {
+							if (onparam.args.length > 0 && onparam.args[0] != null)
+								if (isRestricted(param, uid))
+									onparam.args[0] = new ArrayList<CellInfo>();
+						}
+					}
+				}
+			});
+			Util.log(this, Log.WARN, "Hooked " + on + " uid=" + Process.myUid());
 		} catch (Throwable ex) {
 			Util.bug(this, ex);
 		}
-
-		int uid = Binder.getCallingUid();
-		return getRestricted(context, uid, true);
 	}
 
-	private class XIPhoneStateListener implements IPhoneStateListener {
-		private IPhoneStateListener mListener;
+	@Override
+	protected boolean isRestricted(MethodHookParam param) throws Throwable {
+		int uid = Binder.getCallingUid();
+		return isRestricted(param, uid);
+	}
 
-		public XIPhoneStateListener(IPhoneStateListener listener) {
-			mListener = listener;
+	private boolean isRestricted(MethodHookParam param, int uid) throws Throwable {
+		Context context = null;
+		try {
+			Field fieldContext = findField(param.thisObject.getClass(), "mContext");
+			context = (Context) fieldContext.get(param.thisObject);
+		} catch (Throwable ex) {
+			Util.bug(this, ex);
 		}
-
-		@Override
-		public void onCallForwardingIndicatorChanged(boolean cfi) {
-			mListener.onCallForwardingIndicatorChanged(cfi);
-		}
-
-		@Override
-		public void onCallStateChanged(int state, String incomingNumber) {
-			mListener.onCallStateChanged(state,
-					(String) PrivacyManager.getDefacedProp(Binder.getCallingUid(), "PhoneNumber"));
-		}
-
-		@Override
-		public void onCellInfoChanged(List<CellInfo> cellInfo) {
-			mListener.onCellInfoChanged(new ArrayList<CellInfo>());
-		}
-
-		@Override
-		public void onCellLocationChanged(Bundle location) {
-			mListener.onCellLocationChanged(new Bundle());
-		}
-
-		@Override
-		public void onDataActivity(int direction) {
-			mListener.onDataActivity(direction);
-		}
-
-		@Override
-		public void onDataConnectionStateChanged(int state, int networkType) {
-			mListener.onDataConnectionStateChanged(state, networkType);
-		}
-
-		@Override
-		public void onMessageWaitingIndicatorChanged(boolean mwi) {
-			mListener.onMessageWaitingIndicatorChanged(mwi);
-		}
-
-		@Override
-		public void onServiceStateChanged(ServiceState serviceState) {
-			mListener.onServiceStateChanged(serviceState);
-		}
-
-		@Override
-		public void onSignalStrengthChanged(int asu) {
-			mListener.onSignalStrengthChanged(asu);
-		}
-
-		@Override
-		public void onSignalStrengthsChanged(SignalStrength signalStrength) {
-			mListener.onSignalStrengthsChanged(signalStrength);
-		}
-
-		@Override
-		public void onOtaspChanged(int otaspMode) {
-			mListener.onOtaspChanged(otaspMode);
-		}
-
-		@Override
-		public IBinder asBinder() {
-			return mListener.asBinder();
-		}
+		return getRestricted(context, uid, true);
 	}
 }