+++ CHANGELOG.md
@@ -17,8 +17,9 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Fixed restricting locations for some Android versions ([issue](https://github.com/M66B/XPrivacy/issues/1102))
+* Fixed restricting incoming phone number for some Android versions
 * Fixed removing location and phone state listener
-* Fixed fake value for sim operator name
+* Fixed fake value for SIM (ICC) operator name
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)